add neural emulator tile
I've added a tile for the Neural Emulator, a building related to winning
the game. I've also replaced the art asset for it, making it more in
line with the style of the Nanotech Foundry and Zero Point Energy Plant.

I've created a new type of construction site for the Neural Emulator and
Nanotech Foundry, since it didn't make sense for them to be categorized
as "industry".