implement monolith world-consuming process
The ending of the game's plot involves activating the alien monolith,
which consumes the planet by spreading self-replicating computers.
Showing this process unfolding over time requires some additions beyond
the normal tile projects system, so I've added a new subclass of
TileProject and an alternate version of the world screen used only for
this ending animation.

This commit also moves the current position of the map ui view area to
the World class so that the position will be preserved between screens.