add alternate seed core activation for human victory
in the "good ending", the player allows the AI alignment research
program to happen, which makes it possible  to fail the mission in order
to save the human colonists. instead of the world being consumed by the
alien eschaton, the monolith is reprogrammed to host human minds. this
commit adds alternate versions of the monolith and circuit tiles and
tile projects for activating this alternate ending