make endings part of the quest paths
endings were previously a part of a quest stage, but this meant we
had to include "dummy" quest stages at the end to hold the endings.
now, the quest path "next" field can be either another quest stage
or an ending that is triggered when that path's requirement is met.