refactor predicates
we previously had classes called "TilePredicate" and "WorldPredicate"
that wrapped a predicate with a player-readable description of the
condition that the predicate tests for. this made sense for the tile
project completion requirements, since we need to provide the player
with a list of the requirements.

however, these ended up getting used in places that don't need a string,
in the tile project visibility requirements (which don't appear in the
ui) and in quest path advancement requirements. this commit changes
these places to use plain predicate functions and factories that create
predicate functions. for project completion requirements, which still
need a class to provide a description string, I've switched to using
factories rather than creating a separate subclass for each type of
predicate.