make range input look consistent across browsers
this commit adds css to make the range inputs (sliders) on the settings
page look the same on chrome and firefox, since those two browsers
have very different default appearances for range inputs. it doesn't
work on Internet Explorer or on the older, non-chromium version of
Microsoft Edge, but that's fine because we already didn't support those
browsers. it should also work with the desktop version of Safari but I
haven't tested that because I don't have access to a Mac.

i also added css to disable the dotted line that appears around
buttons when clicked in firefox.