Refactor tile/world predicates to a query system
Having `TilePredicate`s and `WorldPredicate`s in the fields of classes
made comparing or serializing these types a pain. Instead, we replace
these with `TileQuery`s and `WorldQuery`s, which are a simple AST
describing a query. These can then be interpreted into their
corresponding predicates using `queryTile` or `queryWorld`.