move quest hint about moving with keys to an earlier quest
previously there was a quest hint telling the player that the arrow keys
or WASD keys can be used to move around the map when they got to the
"survey the monolith" quest, since the monolith is outside the starting
area and new players won't know how to get there. however, it's possible
that they couldn't get that far without knowing how to move around, for
example if there are no alien ruins randomly placed near the starting
area. to fix this i've moved that hint to the "build xenoarchaeology
lab" quest.