Add serialization schemas for object graph of Game
Currently, every tile subclass needs a schema, and a call to a global
map in `Tile` in order to keep track of the constructors and their
names. In the future, this should be done with decorator metadata.

Also fixes Tile subclass constructors to extend 'typeof Tile' by making
the wasteland variant optional. typeofTileSchema requires this for
TileQuery serialization.