significantly improve wasteland texture variant tracking
if the player creates a construction site and then reverts it to
wasteland, we want to remember which variant of the wasteland texture
was at that position. we were previously doing this with a map within
the world object that could store past texture variants for each
position, but in addition to being inelegant and not fitting with
the actual focus of the World class, this would be extremely difficult
to serialize.

The new solution is to store the variants as a number from 1 to 5,
making them easier to serialize, and then to pass that number
into the construction site tile constructor. the construction site Tiles
can then remember the variant and use it to create a new wasteland tile
with the correct variant.