implement achievements
this adds "achievements" that can be unlocked by doing certain things
in the game. the requirements for unlocking an achievement are evaluated
any time a `Game` checks its quest state, which should be anytime
something significant about the game state
changes.

i've also added a UI page for viewing the list of unlocked achievements
and an option in the settings page to reset achievement progress.