remove EndState field from Game and instead compute it from questStage
this simplifies the way that the state of the story is represented.
we can determine what end state we are in from the current questStage
so there's no need to separately keep track of the endState, and this
simplifies the logic of game.updateQuestState().

The added computation cost is not a problem because we only have to
check the endState when the QuestIndicator finished showing that a quest
has been completed.