change line length limit to 120
when we first set up lining, I set an unconventional line-limit of 140
characters, and then accidentally wrote a commit message claiming it was
120. This commit changes the limit to actually be 120 characters,
and changes code across many files to comply with this limit.

I also had to disable eslint indentation checking for fields because it
was enforcing an incorrect indentation style for wrapped lines.
See the issue at
https://github.com/typescript-eslint/typescript-eslint/issues/1824