make all tile types use texture getter
each tile previously had a texture field (the HTMLImageElement used to
draw that tile on the map canvas), but this was accessed through
getTexture() to support tiles with different variants or that change
appearance based on adjacent tiles. by default this method would just
get the HTMLImageElement from the texture field, but it could be
overloaded.

the problem is tiles that override getTexture() still had to set
a value for that field even though it would never be used. to fix this,
I've removed that field and made getTexture() abstract so that each
Tile subclass has to implement getTexture(), which in most cases just
returns a specific image for that tile type.