refactor Technology to use a multiton
this makes the Technology class a multiton, replacing the previous
system of factory methods.

we need to have a lit of all of the researchable technologies that might
be displayed on the research page. we previously created most
technology instances through the makeUnlockableTechnology factory, which
added those to a global list of unlockable technologies.

By using a multiton, we can get an array that is guaranteed to include
all instances of Technology, and then filter it to get the unlockable
instances. This also simplifies serialization (since we can serialize
just the keys of the instances and don't have to worry about serializing
their recursive fields) and makes it safe to compare Technology
instances by reference.