refactor end-game circuit-expansion process
at the end of the game, the player completed a special tile project that
triggers circuits tiles to spread across the map and consume all other
tiles. the code for this was in EndingWorldScreen, a modified version of
the WorldScreen. i've moved the circuits-expanding code into
MonolithCompletionProject, the TileProject subclass that triggers that
process, and extracted that to a separate file.

I've also cleaned up the control flow for expanding the circuits. I
ignored the linter error that prohibits floating promises, which in my
opinion is worth it to eliminate the convoluted recursion of the old
implementation.