Fix crash after quotes list is exhausted
Since `shuffle` didn't actually allocate a new array, the quotes list
was being emptied as the individual quotes were popped from the sample.
When the sample was supposed to be refilled from the quotes list, both
arrays eventually ended up empty, and they would loop forever as it
counted `undefined` as a repeated quote.